:truck: Restructure persistence layer to avoid ambiguity
