+++ cmd/objekt/main.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/julienschmidt/httprouter"
 	obj_http "github.com/upmahq/objekt/internal/adapter/http"
-	m_repo "github.com/upmahq/objekt/internal/adapter/storage/memory/repository"
-	"github.com/upmahq/objekt/internal/adapter/storage/postgres"
-	p_repo "github.com/upmahq/objekt/internal/adapter/storage/postgres/repository"
+	m_repo "github.com/upmahq/objekt/internal/adapter/persistence/memory/repository"
+	"github.com/upmahq/objekt/internal/adapter/persistence/postgres"
+	p_repo "github.com/upmahq/objekt/internal/adapter/persistence/postgres/repository"
 	"github.com/upmahq/objekt/internal/config"
 	"github.com/upmahq/objekt/internal/core/port"
 	"github.com/upmahq/objekt/internal/core/service"
